+++ iOS/SharedTestUtils/Mocks/ContextualOnboardingPresenterMock.swift
@@ -131,3 +131,66 @@ final class ContextualOnboardingLogicMock: ContextualOnboardingLogic, PrivacyPro
 
     }
 }
+
+// Use to fill parameter list in injection.
+class DummyDaxDialogsManager: DaxDialogsManaging {
+    var isShowingFireDialog: Bool = false
+
+    var shouldShowPrivacyButtonPulse: Bool = false
+
+    var isShowingSearchSuggestions: Bool = false
+
+    var isShowingSitesSuggestions: Bool = false
+
+    var shouldShowFireButtonPulse: Bool = false
+
+    var isAddFavoriteFlow: Bool = false
+
+    var isShowingPrivacyProPromotion: Bool = false
+
+    var privacyProPromotionDialogSeen: Bool = false
+
+    func setTryAnonymousSearchMessageSeen() { }
+
+    func setTryVisitSiteMessageSeen() {}
+
+    func setSearchMessageSeen() {}
+
+    func setFireEducationMessageSeen() {}
+
+    func clearedBrowserData() {}
+
+    func setFinalOnboardingDialogSeen() {}
+
+    func setPrivacyButtonPulseSeen() { }
+
+    func setDaxDialogDismiss() {}
+
+    func enableAddFavoriteFlow() {}
+
+    func isStillOnboarding() -> Bool {
+        false
+    }
+
+    func fireButtonPulseCancelled() {}
+
+    func resumeRegularFlow() {}
+
+    func clearHeldURLData() {}
+
+    func fireButtonPulseStarted() {}
+
+    func nextBrowsingMessageIfShouldShow(for privacyInfo: PrivacyDashboard.PrivacyInfo) -> DuckDuckGo.DaxDialogs.BrowsingSpec? {
+        nil
+    }
+
+    func overrideShownFlagFor(_ spec: DuckDuckGo.DaxDialogs.BrowsingSpec, flag: Bool) {}
+
+    func nextHomeScreenMessageNew() -> DuckDuckGo.DaxDialogs.HomeScreenSpec? {
+        nil
+    }
+
+    func dismiss() {}
+
+    func disableContextualDaxDialogs() {}
+}
